+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Unreleased (??)
 
  - `View` is now a supertrait of `MenuItem`
  - Fixed issues with displaying a slice of menuitems
+ - Added `MenuState`, `MenuBuilder::build_with_state` and `Menu::state`
 
 0.3.1 (2023-08-06)
 ==================